profile using clock cycle

+++ clock_profile/device_kernel.h
@@ -84,7 +84,7 @@ __global__ void initQueues(RingQueue* queues, int** buffers, int cap) {
 template <typename Operator>
 CUTLASS_GLOBAL
 void Kernel(typename Operator::Params params, uint8_t *Signature_Array, 
-            int *Lock_Signature, int *final_sum, int if_split_phase, RingQueue *d_queues, uint8_t *SM_JOBS,
+            int *Lock_Signature, int *final_sum, int if_split_phase, RingQueue *d_queues, int *SM_JOBS,
             int *all_start, int *compute, int *finding, int *checking) {
   // Dynamic shared memory base pointer
   extern __shared__ int SharedStorageBase[];
